Online-Fragebogen

Online und an Testungsterminen abgefragte Personendaten, Konstrukte und aggregierte Spiele-Scores.

data(online)

Format

Data frame mit 91 Zeilen (Teilnehmer) und 76 Variablen:

Testung <date>

TestungsTermin. Die Teilnehmer wurden in 5 Gruppen zu 16—21 Personen aufgeteilt und jede Gruppe führte die Testung an einem eigenen Termin von 09:00—12:00 durch.

ofb <lgl>

Onlineteil bearbeitet. Drei Teilnehmer haben ihren Online-Fragebogen nicht bearbeitet.

p1:p4 <dbl>

Puls. Im Laufe der Testung wurde mehrmals der Puls abgefragt (gezählte Schläge in 10s).

axy <int>

Akzeptanz (x=Spiel, y=Fragennummer), sechsstufige Likertskala

Geschlecht <chr>

Geschlecht (m oder w)

Hand <chr>

Händigkeit (links oder rechts)

SA_Leistung <int>

Selbsteinschätzung der Leistungsfähigkeit (fünfstufige Likertskala).

get_... <dbl>

Gkkt (2—9=verbal, 10—11=numerisch, 12—13=figural)

BIP_LM <dbl>

Leistungsmotivation

BFI

Big 5 <dbl>

CM <dbl>

Competition Motivation

TAI <dbl>

Testängstlichkeit

addifix <dbl>

Numerisches Spiel (Mittlerer Score)

addifix <dbl>

Numerisches Spiel (Mittlerer Score)

snetris <dbl>

Figurales Spiel (Mittlerer Score)

begriffix <dbl>

Verbales Spiel (Mittlerer Score)

References

Rammstedt B, John OP (2005). “Kurzversion des Big Five Inventory (BFI-K).” Diagnostica, 51(4), 195--206. doi: 10.1026/0012-1924.51.4.195 . Hossiep R, Paschen M (2003). Bochumer Inventar zur berufsbezogenen Persönlichkeitsbeschreibung, 2., vollständig überarbeitete Auflage edition. Hogrefe. Wacker A, Jaunzeme J, Jaksztat S (2008). “Eine Kurzform des Prüfungsängstlichkeits-Inventars TAI-G.” Zeitschrift für Pädagogische Psychologie, 22(1), 73--81. doi: 10.1024/1010-0652.22.1.73 . Franken RE, Brown DJ (1995). “Why do people like competition?” Personality and Individual Differences, 19(2), 175--184. Kersting M (2008). “Zur Akzeptanz von Intelligenz- und Leistungstests.” Report Psychologie, 33, 420--433. Moosbrugger H, Oehlschlägel J, Steinwascher M (2011). FAIR-2, Zweite, überarbeitete, ergänzte und normenaktualisierte Auflage des FAIR von Moosbrugger \& Oehlschlägel, 1996 edition. Hans Huber Verlag. Ulfert A, Ott M, Bothe PS, Kersting M (2017). “Gießener kognitiver Kompetenztest (GkKT).” Unveröffentlichtes Manuskript.

See also